+++ crates/wazuh-cert-oauth2-client/src/shared/mod.rs
@@ -1,3 +1,3 @@
 pub mod cli;
 pub mod path;
-pub mod sed_command;
+pub mod replace_tag;